Global rename from Devices to Device
- Update namespace, docs and assemblies.
- Update pipeline yaml and nuspecs accordingly.
- Bump version to 1.3
