+++ .github_changelog_generator
@@ -1,5 +1,5 @@
 user=nanoframework
-project=nanoFramework.Devices.Can
+project=nanoFramework.Device.Can
 issues=true
 add_issues_wo_labels=false
 add_pr_wo_labels=false
