+++ config/filelist.txt
@@ -1,1 +1,1 @@
-**/nanoFramework.Devices.Can.*
+**/nanoFramework.Device.Can.*
