+++ nanoFramework.Device.Can/CanMessageReceivedEventArgs.cs
@@ -3,7 +3,7 @@
 // See LICENSE file in the project root for full license information.
 //
 
-namespace nanoFramework.Devices.Can
+namespace nanoFramework.Device.Can
 {
     /// <summary>
     /// Provides data for the <see cref="CanController.MessageReceived"/> event.
