+++ nanoFramework.Device.Can/packages.config
