--- nanoFramework.Devices.Can.DELIVERABLES.nuspec
@@ -1,32 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2012/06/nuspec.xsd">
-  <metadata>
-    <id>nanoFramework.Devices.Can.DELIVERABLES</id>
-    <version>$version$</version>
-    <title>nanoFramework.Devices.Can.DELIVERABLES</title>
-    <authors>nanoframework</authors>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <releaseNotes>
-    </releaseNotes>
-    <developmentDependency>false</developmentDependency>
-    <projectUrl>https://github.com/nanoframework/lib-nanoFramework.Devices.Can</projectUrl>
-    <iconUrl>https://secure.gravatar.com/avatar/97d0e092247f0716db6d4b47b7d1d1ad</iconUrl>
-    <copyright>Copyright (c) .NET Foundation and Contributors</copyright>
-    <description>** DON'T REFERENCE THIS PACKAGE ** Not meant for development. This package includes the deliverable artifacts of the nanoFramework.Devices.Can assembly for nanoFramework. These are for testing purposes and for updating the native code base of the library.</description>
-    <summary>nanoFramework.Devices.Can.DELIVERABLES is not meant for development.</summary>
-    <tags>
-    </tags>
-    <dependencies>
-      <dependency id="nanoFramework.CoreLibrary" version="1.11.7" />
-      <dependency id="nanoFramework.Runtime.Events" version="1.10.0-preview.1" />
-    </dependencies>
-  </metadata>
-  <files>
-    <file src="nanoFramework.Devices.Can\bin\Release\Stubs\*.*" target="content\Stubs\" />
-    <file src="nanoFramework.Devices.Can\obj\Release\*.txt" target="content\txt\" />
-    <file src="nanoFramework.Devices.Can\obj\Release\*.dump" target="content\dump\" />
-    <file src="nanoFramework.Devices.Can\obj\Release\*.strings" target="content\dump\" />
-    <file src="nanoFramework.Devices.Can\obj\Release\*.resources" target="content\resources\" />
-    <file src="nanoFramework.Devices.Can\obj\Release\*.il" target="content\disasm\" />
-  </files>
-</package>
