--- nanoFramework.Devices.Can.nuspec
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2012/06/nuspec.xsd">
-  <metadata>
-    <id>nanoFramework.Devices.Can</id>
-    <version>$version$</version>
-    <title>nanoFramework.Devices.Can</title>
-    <authors>nanoframework</authors>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <license type="file">LICENSE.md</license>
-    <releaseNotes>
-    </releaseNotes>
-    <readme>docs\README.md</readme>
-    <developmentDependency>false</developmentDependency>
-    <projectUrl>https://github.com/nanoframework/nanoFramework.Devices.Can</projectUrl>
-    <icon>images\nf-logo.png</icon>
-    <repository type="git" url="https://github.com/nanoframework/nanoFramework.Devices.Can" commit="$commit$" />
-    <copyright>Copyright (c) .NET Foundation and Contributors</copyright>
-    <description>This package includes the nanoFramework.Devices.Can assembly for .NET nanoFramework C# projects.
-This package requires a target with nanoFramework.Devices.Can v$nativeVersion$ (checksum $checksum$).</description>
-    <tags>nanoFramework C# csharp netmf netnf canbus can</tags>
-    <dependencies>
-      <dependency id="nanoFramework.CoreLibrary" version="1.12.0" />
-      <dependency id="nanoFramework.Runtime.Events" version="1.10.0" />
-    </dependencies>
-  </metadata>
-  <files>
-    <file src="nanoFramework.Devices.Can\bin\Release\nanoFramework.Devices.Can.dll" target="lib\nanoFramework.Devices.Can.dll" />
-    <file src="nanoFramework.Devices.Can\bin\Release\nanoFramework.Devices.Can.pdb" target="lib\nanoFramework.Devices.Can.pdb" />
-    <file src="nanoFramework.Devices.Can\bin\Release\nanoFramework.Devices.Can.pdbx" target="lib\nanoFramework.Devices.Can.pdbx" />
-    <file src="nanoFramework.Devices.Can\bin\Release\nanoFramework.Devices.Can.pe" target="lib\nanoFramework.Devices.Can.pe" />
-    <file src="nanoFramework.Devices.Can\bin\Release\nanoFramework.Devices.Can.xml" target="lib\nanoFramework.Devices.Can.xml" />
-    <!-- readme -->
-    <file src="assets\readme.txt" target="" />
-    <file src="README.md" target="docs\" />
-    <file src="assets\nf-logo.png" target="images" />
-    <file src="LICENSE.md" target="" />
-  </files>
-</package>
