+++ version.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://raw.githubusercontent.com/dotnet/Nerdbank.GitVersioning/master/src/NerdBank.GitVersioning/version.schema.json",
-  "version": "1.2.4",
+  "version": "1.3",
   "assemblyVersion": {
     "precision": "revision"
   },
